@@ -21,8 +21,8 @@
 
       <PublishRepositoryUrl>true</PublishRepositoryUrl>
 
-      <TargetFramework>net6.0</TargetFramework>
-      <LangVersion>10.0</LangVersion>
+      <TargetFramework>net8.0</TargetFramework>
+      <LangVersion>12.0</LangVersion>
       <Nullable>enable</Nullable>
    </PropertyGroup>
 
