@@ -82,8 +82,7 @@ public class UInt160 : UIntBase, IComparable<UInt160>, IEquatable<UInt160>
    /// </summary>
    public static new UInt160 Parse(string value)
    {
-      if (value == null)
-         throw new ArgumentNullException(nameof(value));
+      ArgumentNullException.ThrowIfNull(value);
       if (value.StartsWith("0x", StringComparison.InvariantCultureIgnoreCase))
          value = value.Substring(2);
       if (value.Length != LENGTH * 2)
