@@ -12,10 +12,7 @@ public class BackgroundTaskQueue : IBackgroundTaskQueue
 
    public void QueueBackgroundWorkItem(Func<CancellationToken, Task> workItem)
    {
-      if (workItem == null)
-      {
-         throw new ArgumentNullException(nameof(workItem));
-      }
+      ArgumentNullException.ThrowIfNull(workItem);
 
       _workItems.Enqueue(workItem);
       _signal.Release();
