@@ -11,23 +11,19 @@ namespace MithrilShards.Chain.Bitcoin.Consensus;
 public class InitialBlockDownloadTracker : IInitialBlockDownloadTracker
 {
    readonly ILogger<InitialBlockDownloadTracker> _logger;
-   readonly IEventBus _eventBus;
    readonly IChainState _chainState;
    readonly IConsensusParameters _consensusParameters;
    readonly IDateTimeProvider _dateTimeProvider;
    private readonly Target? _minimumChainWork;
    private readonly long _maxTipAge;
-   readonly EventSubscriptionManager _subscriptionManager = new();
 
    public InitialBlockDownloadTracker(ILogger<InitialBlockDownloadTracker> logger,
-                                      IEventBus eventBus,
                                       IChainState chainState,
                                       IConsensusParameters consensusParameters,
                                       IOptions<BitcoinSettings> options,
                                       IDateTimeProvider dateTimeProvider)
    {
       _logger = logger;
-      _eventBus = eventBus;
       _chainState = chainState;
       _consensusParameters = consensusParameters;
       _dateTimeProvider = dateTimeProvider;
@@ -38,7 +34,7 @@ public class InitialBlockDownloadTracker : IInitialBlockDownloadTracker
       _minimumChainWork = options.Value.MinimumChainWork ?? _consensusParameters.MinimumChainWork;
       if (_minimumChainWork < _consensusParameters.MinimumChainWork)
       {
-         _logger.LogWarning($"{nameof(_minimumChainWork)} set below default value of {_consensusParameters.MinimumChainWork}");
+         _logger.LogWarning("_minimumChainWork set below default value of {DefaultMinimumChainWork}", _consensusParameters.MinimumChainWork);
       }
 
       _maxTipAge = options.Value.MaxTipAge;
