@@ -65,7 +65,7 @@ public partial class PingPongProcessor : BaseProcessor,
          ping.Nonce = _randomNumberGenerator.GetUint64();
       }
 
-      await SendMessageAsync(ping).ConfigureAwait(false);
+      await SendMessageAsync(ping, cancellationToken).ConfigureAwait(false);
 
       _status.PingSent(_dateTimeProvider.GetTimeMicros(), ping);
       logger.LogDebug("Sent ping request with nonce {PingNonce}", _status.PingRequestNonce);
@@ -86,24 +86,24 @@ public partial class PingPongProcessor : BaseProcessor,
    async ValueTask<bool> INetworkMessageHandler<PingMessage>.ProcessMessageAsync(PingMessage message, CancellationToken cancellation)
    {
       logger.LogDebug("Received ping with nonce {PingNonce}.", message.Nonce);
-      await SendMessageAsync(new PongMessage { Nonce = message.Nonce }).ConfigureAwait(false);
+      await SendMessageAsync(new PongMessage { Nonce = message.Nonce }, cancellation).ConfigureAwait(false);
 
       return true;
    }
 
-   ValueTask<bool> INetworkMessageHandler<PongMessage>.ProcessMessageAsync(PongMessage message, CancellationToken cancellation)
+   async ValueTask<bool> INetworkMessageHandler<PongMessage>.ProcessMessageAsync(PongMessage message, CancellationToken cancellation)
    {
       if (_status.PingRequestNonce != 0 && message.Nonce == _status.PingRequestNonce)
       {
          (ulong Nonce, long RoundTrip) = _status.PongReceived(_dateTimeProvider.GetTimeMicros());
          logger.LogDebug("Received pong with nonce {PingNonce} in {PingRoundTrip} usec.", Nonce, RoundTrip);
-         _pingCancellationTokenSource.Cancel();
+         await _pingCancellationTokenSource.CancelAsync().ConfigureAwait(false);
       }
       else
       {
          logger.LogDebug("Received pong with wrong nonce: {PingNonce}", _status.PingRequestNonce);
       }
 
-      return new ValueTask<bool>(true);
+      return true;
    }
 }
