@@ -16,117 +16,99 @@ using MithrilShards.Core.Shards;
 
 namespace MithrilShards.Network.Bedrock;
 
-public class BedrockNetworkShard : IMithrilShard
+public class BedrockNetworkShard(
+   ILogger<BedrockNetworkShard> logger,
+   IEnumerable<IServerPeerConnectionGuard> serverPeerConnectionGuards,
+   IOptions<ForgeConnectivitySettings> settings,
+   IServiceProvider serviceProvider
+   ) : IMithrilShard
 {
-   readonly ILogger<BedrockNetworkShard> _logger;
-   readonly IEventBus _eventBus;
-   private readonly IEnumerable<IServerPeerConnectionGuard> _serverPeerConnectionGuards;
-   private readonly IServiceProvider _serviceProvider;
-   private readonly ForgeConnectivitySettings _settings;
-   private readonly List<Server> _serverPeers;
+   private readonly ForgeConnectivitySettings _settings = settings?.Value ?? throw new ArgumentNullException(nameof(settings));
+   private readonly List<Server> _serverPeers = [];
 
-   public BedrockNetworkShard(ILogger<BedrockNetworkShard> logger,
-                             IEventBus eventBus,
-                             IEnumerable<IServerPeerConnectionGuard> serverPeerConnectionGuards,
-                             IOptions<ForgeConnectivitySettings> settings,
-                             IServiceProvider serviceProvider)
-   {
-      _logger = logger;
-      _eventBus = eventBus;
-      _serverPeerConnectionGuards = serverPeerConnectionGuards;
-      _serviceProvider = serviceProvider;
-      _settings = settings?.Value ?? throw new ArgumentNullException(nameof(settings));
-      _serverPeers = new List<Server>();
-   }
-
-   public ValueTask InitializeAsync(CancellationToken cancellationToken)
+   public Task InitializeAsync(CancellationToken cancellationToken)
    {
       CreateServerInstances();
-      return default;
+      return Task.CompletedTask;
    }
 
-   public ValueTask StartAsync(CancellationToken cancellationToken)
+   public Task StartAsync(CancellationToken cancellationToken)
    {
       foreach (Server serverPeer in _serverPeers)
       {
          _ = serverPeer.StartAsync(cancellationToken);
       }
 
-      return default;
+      return Task.CompletedTask;
    }
 
-   public ValueTask StopAsync(CancellationToken cancellationToken)
+   public async Task StopAsync(CancellationToken cancellationToken)
    {
-      foreach (Server serverPeer in _serverPeers)
-      {
-         _ = serverPeer.StopAsync();
-      }
+      var tasks = _serverPeers.Select(serverPeer => serverPeer.StopAsync(cancellationToken)).ToArray();
 
-      return default;
+      await Task.WhenAll(tasks).ConfigureAwait(false);
    }
 
    private void CreateServerInstances()
    {
-      using (_logger.BeginScope("CreateServerInstances"))
+      using var _ = logger.BeginScope("CreateServerInstances");
+      logger.LogInformation("Loading Forge Server listeners configuration.");
+
+      if (serverPeerConnectionGuards.Any())
       {
-         _logger.LogInformation("Loading Forge Server listeners configuration.");
-
-         if (_serverPeerConnectionGuards.Any())
-         {
-            _logger.LogInformation(
-               "Using {PeerConnectionGuardsCount} peer connection guards: {PeerConnectionGuards}.",
-               _serverPeerConnectionGuards.Count(),
-               _serverPeerConnectionGuards.Select(guard => guard.GetType().Name)
-               );
-         }
-         else
-         {
-            _logger.LogWarning("No peer connection guards detected.");
-         }
-
-         if (_settings.Listeners?.Count > 0)
-         {
-            _logger.LogInformation("Found {ConfiguredListeners} listeners in configuration.", _serverPeers.Count);
-
-            ServerBuilder builder = new ServerBuilder(_serviceProvider)
-               .UseSockets(sockets =>
-               {
-                  sockets.Options.NoDelay = true;
-
-                  foreach (ServerPeerBinding binding in _settings.Listeners)
-                  {
-                     IPEndPoint localEndPoint = binding.GetIPEndPoint();
-
-                     if (!binding.HasPublicEndPoint())
-                     {
-                        binding.PublicEndPoint = new IPEndPoint(IPAddress.Loopback, localEndPoint.Port).ToString();
-                     }
-
-                     binding.TryGetPublicIPEndPoint(out IPEndPoint? publicEndPoint);
-
-                     _logger.LogInformation("Added listener to local endpoint {ListenerLocalEndpoint}. (remote {ListenerPublicEndpoint})", localEndPoint, publicEndPoint);
-
-                     sockets.Listen(
-                        localEndPoint.Address,
-                        localEndPoint.Port,
-                        builder => builder
-                           .UseConnectionLogging()
-                           .UseConnectionHandler<MithrilForgeServerConnectionHandler>()
-                        );
-                  }
-               }
+         logger.LogInformation(
+            "Using {PeerConnectionGuardsCount} peer connection guards: {PeerConnectionGuards}.",
+            serverPeerConnectionGuards.Count(),
+            serverPeerConnectionGuards.Select(guard => guard.GetType().Name)
             );
+      }
+      else
+      {
+         logger.LogWarning("No peer connection guards detected.");
+      }
 
-            builder.ShutdownTimeout = TimeSpan.FromSeconds(_settings.ForceShutdownAfter);
+      if (_settings.Listeners?.Count > 0)
+      {
+         logger.LogInformation("Found {ConfiguredListeners} listeners in configuration.", _serverPeers.Count);
 
-            Server server = builder.Build();
+         ServerBuilder builder = new ServerBuilder(serviceProvider)
+            .UseSockets(sockets =>
+            {
+               sockets.Options.NoDelay = true;
 
-            _serverPeers.Add(server);
-         }
-         else
-         {
-            _logger.LogWarning("No binding information found in configuration file, no Forge Servers available.");
-         }
+               foreach (ServerPeerBinding binding in _settings.Listeners)
+               {
+                  IPEndPoint localEndPoint = binding.GetIPEndPoint();
+
+                  if (!binding.HasPublicEndPoint())
+                  {
+                     binding.PublicEndPoint = new IPEndPoint(IPAddress.Loopback, localEndPoint.Port).ToString();
+                  }
+
+                  binding.TryGetPublicIPEndPoint(out IPEndPoint? publicEndPoint);
+
+                  logger.LogInformation("Added listener to local endpoint {ListenerLocalEndpoint}. (remote {ListenerPublicEndpoint})", localEndPoint, publicEndPoint);
+
+                  sockets.Listen(
+                     localEndPoint.Address,
+                     localEndPoint.Port,
+                     builder => builder
+                        .UseConnectionLogging()
+                        .UseConnectionHandler<MithrilForgeServerConnectionHandler>()
+                     );
+               }
+            }
+         );
+
+         builder.ShutdownTimeout = TimeSpan.FromSeconds(_settings.ForceShutdownAfter);
+
+         Server server = builder.Build();
+
+         _serverPeers.Add(server);
+      }
+      else
+      {
+         logger.LogWarning("No binding information found in configuration file, no Forge Servers available.");
       }
    }
 }
